refactor: move EmojiEnum to non-type file
